Date spinner and date picker dialog
Date picker dialog can be opened by selecting a spinner item, and selected
date is then added to the spinner.

Next steps

-  On clicking "Add Item", retrieve the selected date (if any)
from the spinner, convert it to another date representation, and add it
to the TodoItem that will be saved in the database.
- Adapt the customised ArrayAdapter for the ListView so that the date of
a TodoItem is displayed in the ListView item.