Add highlighting of selected ListView items, and title of contextual action bar
Use of a selector (res/drawable/list_item_selector.xml)  applied to the
background of the ListView items. Different background colour for the state
android:state_activated. This state is set to true when an item is selected
through the contextual action mode.

The title of the contextual action bar can be set through the ActionMode
object received in the callbacks of the MultiChoiceModeListener.